+++ src/main/java/ru/yandex/practicum/filmorate/mapper/UserLogRowMapper.java
@@ -13,8 +13,8 @@ public class UserLogRowMapper implements RowMapper<UserLog> {
     public UserLog mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         UserLog userLog = new UserLog();
         userLog.setTimeStamp(resultSet.getTimestamp("action_timestamp").toLocalDateTime());
-        userLog.setUserId(resultSet.getLong("user_id"));
-        userLog.setEntityId(resultSet.getLong("entity_id"));
+        userLog.setUserId(resultSet.getInt("user_id"));
+        userLog.setEntityId(resultSet.getInt("entity_id"));
         userLog.setEventType(resultSet.getString("event_type"));
         userLog.setOperation(resultSet.getString("operation"));;
         return userLog;
